Build gem with GitHub Actions
Takes inspiration from https://github.com/alphagov/rubocop-govuk/pull/45
I've saved the API credential under 'Secrets' in this repo, as
'GEM_HOST_API_KEY', which is referenced by the GitHub Action.

The logic is conditional so that a gem is only built and published
if the version doesn't match the latest version on rubygems.
As we haven't been able to release 0.11.1 yet, this should mean
that gem is released when this PR is merged.