Remove Jenkins integration
GOV.UK is intended to retire the use of Jenkins CI for rubygems and this
project is one of the last ones using the GOV.UK Jenkins.

This project is not owned by GOV.UK, as it is part of the tech-docs
tools, so it was peculiar that it used GOV.UK's CI. As it doesn't
actually have any tests I didn't want to set it up to a new CI.

Ideally we should see if we see need this fork and whether we can adopt
the regular middleman-search gem.